backend/requirements.txt
Flask>=2.1.2 
PyJWT>=1.7.1 
python-dotenv>=0.20.0 
PyMySQL>=1.0.2
Werkzeug>=2.1.2